pyproject.toml
[tool.poetry]
name = "lapis"
version = "0.1.0"
description = ""
authors = ["Herman <111755919+MrBenoit@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
disnake = "^2.9.1"
pillow = "^10.2.0"
openai = "^1.9.0"
asyncpg = "^0.29.0"
qrcode = "^7.4.2"
black = "^23.12.1"
langdetect = "^1.0.9"
python-dotenv = "^1.0.1"
lavalink = "^5.1.0"
matrix = "^3.0.0"
sqlalchemy = "^2.0.25"
sengo = "^0.0.2"
segno = "^1.6.0"
pylint = "^3.0.3"
mafic = "^2.10.0"
requests = "^2.31.0"
somerandomapi = "^0.0.6"
emoji = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
